Fix statistics
